backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
starknet-py==0.23.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
aiosmtplib==3.0.1
jinja2==3.1.2
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
statsmodels==0.14.0
python-multipart==0.0.6
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2
pytz==2023.3

